+++ NEWS.md
@@ -6,8 +6,10 @@ editor_options:
 
 # gginnards 0.1.1
 
--   Update `geom_debug()` to avoid some spurious warnings triggered by
-    unrecognised arguments and aesthetics.
+-   Update `geom_debug()` to avoid spurious warnings triggered by
+    unrecognised arguments and aesthetics. Aesthetics silently accepted
+    include all those used by geoms from packages 'ggplot2', 'ggpp', and
+    'ggrepel'.
 
 # gginnards 0.1.0-1